+++ levior/filters/_helpers.py
@@ -1,6 +1,13 @@
 from trimgmi import Line, LineType
 
 
+def str_to_linetype(type_s: str) -> LineType:
+    try:
+        return getattr(LineType, type_s.upper())
+    except AttributeError:
+        return None
+
+
 def is_text(line: Line):
     return line.type in [LineType.REGULAR,
                          LineType.QUOTE,